Create test2
are given a string S which represents a date in the format YYYYMMDD. You are also given an integer N.
Write a Python program using Pandas to print all the N dates from the given date in the format YYYY-MM-DD.

Function Description
In the provided code snippet, implement the provided dateSeries(N, S) method using the variables to print all the N dates from the given date. You can write your code in the space below the phra|
HERE”,

‘There will be multiple test cases running so the Input and Output should match exactly as provided,
‘The base Output variable result is set to a default value of -404 which can be modified. Additionally, you can add or remove these output variables.

Input Format
‘The first line of the input contains an integer, denoting the value of
‘The second line of the input contains a string, denoting the value of S.

‘Sample Input
2
20130101

Output Format
‘The output contains Nspace-separated strings, denoting all the N dates from the given date.

‘Sample Output
2013-01-01 2013-01-02

Explanation
= "20130101" and N=2.

   

All the N dates from the given date in the format YYYY-MM-DD would be '2013-01-01" and ‘2013-01-02. solution